♻ Dto 클래스 convertToxx 메서드 삭제
- Util 패키지에 Mapper 클래스를 추가한다.
- Mapper 클래스는 ObjectMapper 를 사용하여 Dto 인스턴스를 Domain 인스턴스로 변환해주는 convertToDoamin 메서드가 존재한다.
- 제네릭으로 구현되어 있으며 반환 타입, 변환 타입, 인자 객체 타입 모두 자유롭게 지정이 가능하다.

+++ src/test/java/com/kakao/cafe/integration/ArticleControllerTest.java
@@ -22,6 +22,7 @@ import org.springframework.util.MultiValueMap;
 
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 import java.util.Vector;
@@ -44,7 +45,7 @@ public class ArticleControllerTest {
     MockMvc mvc;
 
     @Autowired
-    DomainRepository<Article, Long> repository;
+    DomainRepository<Article, Integer> repository;
 
     static List<Article> articles = new Vector<>();
